Correct rate change callback parameters

+++ examples/wireless/rate-adaptation-distance.cc
@@ -137,7 +137,7 @@ NodeStatistics::GetDatafile ()
 }
 
 
-void RateCallback (std::string path, uint32_t rate, Mac48Address dest)
+void RateCallback (std::string path, uint64_t rate, Mac48Address dest)
 {
   NS_LOG_INFO ((Simulator::Now ()).GetSeconds () << " " << dest << " Rate " <<  rate);
 }
